refactoring geomechanics solvers
